Remember that the alignment in GLSL is different with C/C++, so we really need to know the structures' alignment in advances.
That's why we cannot copy and write the buffer directly for most of the structures.

So I change the SetBlockValue interface to global function SetCameraMatricesBlock,
because so far I think we should always know what kind of UniformBlock we want to set so that we should also know its structures and alignment.