Now the OpenGL shading can use UniformBlock in the Materials.
Rememeber that we need to setup the binding point between the UniformBlock in different shader and the Buffer Object.

For example, if we bind the CameraMatrix UniformBlock in 2 different shaders to the same binding point,
and we also bind a Buffer Object which contains the matrices of the CameraMatrix,
then the 2 shaders can share the same data in the same Buffer through the binding point.

And, if we have a different UniformBlock called SpecularLight, then we should use a different binding point,
and also bind another Buffer Object to the new binding point. Then we can set the data of SpecularLight correctly.